% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{lookup}
\alias{lookup}
\title{Lookup a subset searchable list}
\usage{
lookup(...)
}
\arguments{
\item{slist}{an object of class Rsearchable}

\item{criteria}{a list of criteria, with name equal to a Rseachable field name and value equal to value}

\item{fun}{a function to compare criteria names with values. Default \code{==}.}
}
\description{
Return a subset of searchable list according to a set of criteria
}
\seealso{
\code{\link{Rsearchable}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{lists,}
\keyword{searchable}
