% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{names.searchable}
\alias{names.searchable}
\title{Names}
\usage{
\method{names}{searchable}(slist)
}
\arguments{
\item{obj}{an object of class Rsearchable}
}
\description{
get or set the names of an object of class Rsearchable
}
\seealso{
\code{\link{Rsearchable}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{lists,}
\keyword{searchable}
