% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\docType{class}
\name{searchable-class}
\alias{searchable-class}
\alias{searchable}
\title{A class of searchable lists}
\arguments{
\item{obj}{the input list to transform in a searchable list#'}
}
\description{
Define a searchable list object
}
\seealso{
\code{\link{Rsearchable}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{lists,}
\keyword{searchable}
