% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{select}
\alias{select}
\title{Select a subset searchable list}
\usage{
select(...)
}
\arguments{
\item{slist}{an object of class Rsearchable}

\item{field}{the name of the field to select the children}

\item{values}{a list of values to match.}
}
\description{
Return a subset of searchable list according to values of a give field
}
\seealso{
\code{\link{Rsearchable}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{lists,}
\keyword{searchable}
